trying to fix GPU usage

+++ lstm_tf_imdb_modified.py
@@ -101,7 +101,7 @@ class LSTM_Model(object):
                 word_embedding = tf.get_variable('word_embedding', shape=[vocabulary_size, dim_proj],
                                                   initializer=tf.constant_initializer(random_embedding),dtype=tf.float32)
 
-                unrolled_inputs=tf.reshape(self._inputs,[1,-1])
+                unrolled_inputs=tf.reshape(self._inputs,[1,-1],dtype=tf.int64)
                 embedded_inputs = tf.nn.embedding_lookup(word_embedding, unrolled_inputs)
                 embedded_inputs = tf.reshape(embedded_inputs, [MAXLEN, BATCH_SIZE, dim_proj])
